+++ packages/harpstrata/src/harpstrata/Apparatus/constants/WILDE_TUNED_APPARATUS.ts
@@ -7,7 +7,7 @@ import type { HalfstepIndexMatrix } from '../HalfstepIndex'
 // prettier-ignore
 const halfstepIndexMatrix: HalfstepIndexMatrix = [
   //    1          2          3          4          5          6          7          8          9         10
-  [ 3        , undefined, undefined, 15       , 18       , undefined, undefined, 25       , undefined, undefined ],
+  [ 3        , undefined, undefined, 15       , undefined, undefined, undefined, 27       , 32       , undefined ],
   [ 0        , 4        , 7        , 12       , 16       , 16       , 19       , 24       , 28       , 33        ],
   [ 2        , 7        , 11       , 14       , 17       , 19       , 23       , 26       , 31       , 36        ],
   [ 1        , 6        , 10       , 13       , undefined, 18       , 22       , 25       , 30       , 35        ],
@@ -18,7 +18,7 @@ const halfstepIndexMatrix: HalfstepIndexMatrix = [
 // prettier-ignore
 const interactionMatrix: InteractionMatrix = [
   //    1          2          3          4          5          6          7          8          9         10
-  [ OVERBLOW1, undefined, undefined, OVERBLOW1, OVERBLOW1, undefined, undefined, OVERBLOW1, undefined, undefined ],
+  [ OVERBLOW1, undefined, undefined, OVERBLOW1, undefined, undefined, undefined, OVERBLOW1, OVERBLOW1, undefined ],
   [ BLOW     , BLOW     , BLOW     , BLOW     , BLOW     , BLOW     , BLOW     , BLOW     , BLOW     , BLOW      ],
   [ DRAW     , DRAW     , DRAW     , DRAW     , DRAW     , DRAW     , DRAW     , DRAW     , DRAW     , DRAW      ],
   [ BEND1    , BEND1    , BEND1    , BEND1    , undefined, BEND1    , BEND1    , BEND1    , BEND1    , BEND1     ],