+++ packages/harpguru-core/src/styles/default-colorscheme/default-colorscheme.ts
@@ -18,14 +18,16 @@ const degreeColors = {
 } as const
 
 const pageColor = '#fefefe' as const
-const homeRowsColor = '#eee' as const
-const inertOutline = '#aaa' as const
-const holeNumbersColor = '#495057' as const
+const homeRowsColor = '#ddd' as const
+const inertOutline = '#555' as const
+const activeOutline = 'black' as const
+const holeNumbersColor = inertOutline
 
 export const colors: ColorSheme = {
   pageColor,
   degreeColors,
   homeRowsColor,
   inertOutline,
+  activeOutline,
   holeNumbersColor,
 } as const