+++ packages/harpguru-core/src/utils/get-menu-styles-and-animation-vals/get-menu-styles-and-animation-vals.ts
@@ -5,10 +5,12 @@ import {
   sub,
   multiply,
   add,
-  Node,
+  divide,
+  Extrapolate,
 } from 'react-native-reanimated'
+import type { Node } from 'react-native-reanimated'
 import { StyleSheet, Dimensions } from 'react-native'
-import type { TextStyle, ViewStyle } from 'react-native'
+import type { ViewStyle } from 'react-native'
 
 import { getSizes, colors } from '../../styles'
 
@@ -16,9 +18,7 @@ type MenuStyles = {
   readonly animated: ViewStyle
   readonly overlay: ViewStyle
   readonly mainContents: ViewStyle
-  readonly rotatedLabel: ViewStyle
-  readonly labelAligner: ViewStyle
-  readonly text: TextStyle
+  readonly label: ViewStyle
 }
 
 type StyleAndAnimationVals = {
@@ -27,7 +27,7 @@ type StyleAndAnimationVals = {
   readonly menuSlideYTranslation: Node<number>
   readonly menuScale: Node<number>
   readonly menuBackgroundColor: Node<number>
-  readonly labelOpacity: Node<number>
+  readonly menuOpacity: Node<number>
   readonly labelCounterScale: Node<number>
 }
 
@@ -37,10 +37,24 @@ export const getMenuStylesAndAnimationVals = (
   stashDirection: 'TOP' | 'BOTTOM'
 ): StyleAndAnimationVals => {
   const sizes = getSizes()
-  const { labelProtrusion, 9: fontSize, 7: borderRadius } = sizes
+  const {
+    labelProtrusion: unscaledLabelProtrusion,
+    9: borderRadius,
+    overlayOpacity,
+  } = sizes
   const outwardXMultiplier = 1
   const outwardYMultiplier = stashDirection === 'TOP' ? -1 : 1
-  const labelRotation = '90deg'
+
+  const { width: windowWidth, height: windowHeight } = Dimensions.get('window')
+  const deviceShortSide =
+    windowWidth < windowHeight ? windowWidth : windowHeight
+  const deviceLongSide = windowWidth > windowHeight ? windowWidth : windowHeight
+
+  const menuHiddenScale = 0.2 // 0.5 would have both tabs fill exactly half the screen height
+  const menuHiddenYOffsetFactor = 0.3
+  const menuScaleTranslationFactor = (1 - menuHiddenScale) / 2
+
+  const labelProtrusion = unscaledLabelProtrusion / menuHiddenScale
 
   const styles = StyleSheet.create({
     animated: {
@@ -53,48 +67,28 @@ export const getMenuStylesAndAnimationVals = (
       flexDirection: 'row',
       justifyContent: 'flex-start',
       borderRadius,
-      opacity: 0.7,
     },
     mainContents: {
       ...StyleSheet.absoluteFillObject,
       flexDirection: 'row',
       left: labelProtrusion,
     },
-    rotatedLabel: {
-      overflow: 'visible',
-      alignSelf: 'center',
+    label: {
       alignItems: 'center',
       justifyContent: 'center',
-      height: labelProtrusion,
       width: labelProtrusion,
-      transform: [{ rotate: labelRotation }],
-    },
-    labelAligner: {
-      alignItems: 'center',
-      width: 500,
-    },
-    text: {
-      fontSize,
     },
   })
-
-  const menuHiddenScale = 0.49 // 0.5 would have both tabs fill exactly half the screen height
-  const menuScaleTranslationFactor = (1 - menuHiddenScale) / 2
 
   // Animation values
   const hideMenuVal = useTimingTransition(hideMenu, {
-    duration: 400,
+    duration: 300,
     easing: Easing.inOut(Easing.ease),
   })
   const hideLabelVal = useTimingTransition(hideLabel, {
-    duration: 400,
+    duration: 300,
     easing: Easing.inOut(Easing.ease),
   })
-
-  const { width: windowWidth, height: windowHeight } = Dimensions.get('window')
-  const deviceShortSide =
-    windowWidth < windowHeight ? windowWidth : windowHeight
-  const deviceLongSide = windowWidth > windowHeight ? windowWidth : windowHeight
 
   // Menu animation values
   const hideMenuXTranslation = interpolate(hideMenuVal, {
@@ -114,7 +108,11 @@ export const getMenuStylesAndAnimationVals = (
     inputRange: [0, 1],
     outputRange: [
       0,
-      multiply(deviceShortSide, menuScaleTranslationFactor, outwardYMultiplier),
+      multiply(
+        multiply(deviceShortSide, menuHiddenYOffsetFactor),
+        menuScaleTranslationFactor,
+        outwardYMultiplier
+      ),
     ],
   })
   const hideLabelTranslation = interpolate(hideLabelVal, {
@@ -131,15 +129,16 @@ export const getMenuStylesAndAnimationVals = (
     inputRange: [0, 1],
     outputRange: [colors.pageColor, colors.homeRowsColor],
   })
+  const menuOpacity = interpolate(hideMenuVal, {
+    inputRange: [0, 1],
+    outputRange: [overlayOpacity, 1],
+  })
 
   // Label animation values
-  const labelOpacity = interpolate(hideMenuVal, {
-    inputRange: [0, 1],
-    outputRange: [0, 1],
-  })
   const labelCounterScale = interpolate(menuScale, {
-    inputRange: [menuHiddenScale, 1],
-    outputRange: [1, menuHiddenScale],
+    inputRange: [menuHiddenScale, 0.4],
+    outputRange: [divide(1, menuHiddenScale), 0],
+    extrapolate: Extrapolate.CLAMP,
   })
 
   return {
@@ -148,7 +147,7 @@ export const getMenuStylesAndAnimationVals = (
     menuSlideYTranslation,
     menuScale,
     menuBackgroundColor,
-    labelOpacity,
+    menuOpacity,
     labelCounterScale,
   }
 }